honksquad: refactor: move Self-Aware trait into RussStation fork dir
