+++ Content.Shared/RussStation/Traits/SelfAwareComponent.cs
@@ -1,11 +1,9 @@
 using Robust.Shared.GameStates;
 
-namespace Content.Shared.Traits.Assorted;
+namespace Content.Shared.RussStation.Traits;
 
 /// <summary>
 /// An entity with this component can see exact damage numbers when examining themselves.
 /// </summary>
 [RegisterComponent, NetworkedComponent]
-public sealed partial class SelfAwareComponent : Component
-{
-}
+public sealed partial class SelfAwareComponent : Component;
